move AI files to the AI directory
