+++ src/hexGame/hexGame/Player.cpp
@@ -1,7 +1,6 @@
 
 #include "Player.hpp"
 #include "Human.hpp"
-#include "RandomAI.hpp"
 #include "Game.hpp"
 
 Player::Player(Color color): m_color(color) {}
